Support SpellCheck in auto tests.
