--- BingSearchv7/BingSpellCheck/bing-search-dotnet.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26730.12
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "bing-search-dotnet", "bing-search-dotnet.csproj", "{1BAB52EF-578B-44C0-9708-F9E7EFAC3DD5}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1BAB52EF-578B-44C0-9708-F9E7EFAC3DD5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1BAB52EF-578B-44C0-9708-F9E7EFAC3DD5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1BAB52EF-578B-44C0-9708-F9E7EFAC3DD5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1BAB52EF-578B-44C0-9708-F9E7EFAC3DD5}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {23A7C819-7D1E-4C8E-8863-7190AC370D3E}
-	EndGlobalSection
-EndGlobal
